moved the default provider for standardized domain parameters into its own class
